more moving legacy migrations
