+++ packages/tlschema/src/shapes/TLArrowShape.ts
@@ -1,4 +1,3 @@
-import { defineMigrations } from '@tldraw/store'
 import { T } from '@tldraw/validate'
 import { vecModelValidator } from '../misc/geometry-types'
 import { StyleProp } from '../styles/StyleProp'
@@ -77,107 +76,3 @@ export type TLArrowShapeProps = ShapePropsType<typeof arrowShapeProps>
 
 /** @public */
 export type TLArrowShape = TLBaseShape<'arrow', TLArrowShapeProps>
-
-export const ArrowMigrationVersions = {
-	AddLabelColor: 1,
-	AddIsPrecise: 2,
-	AddLabelPosition: 3,
-} as const
-
-/** @internal */
-// eslint-disable-next-line deprecation/deprecation
-export const arrowShapeMigrations = defineMigrations({
-	currentVersion: ArrowMigrationVersions.AddLabelPosition,
-	migrators: {
-		[ArrowMigrationVersions.AddLabelColor]: {
-			up: (record) => {
-				return {
-					...record,
-					props: {
-						...record.props,
-						labelColor: 'black',
-					},
-				}
-			},
-			down: (record) => {
-				const { labelColor: _, ...props } = record.props
-				return {
-					...record,
-					props,
-				}
-			},
-		},
-
-		[ArrowMigrationVersions.AddIsPrecise]: {
-			up: (record) => {
-				const { start, end } = record.props
-				return {
-					...record,
-					props: {
-						...record.props,
-						start:
-							(start as TLArrowShapeTerminal).type === 'binding'
-								? {
-										...start,
-										isPrecise: !(
-											start.normalizedAnchor.x === 0.5 && start.normalizedAnchor.y === 0.5
-										),
-									}
-								: start,
-						end:
-							(end as TLArrowShapeTerminal).type === 'binding'
-								? {
-										...end,
-										isPrecise: !(end.normalizedAnchor.x === 0.5 && end.normalizedAnchor.y === 0.5),
-									}
-								: end,
-					},
-				}
-			},
-			down: (record: any) => {
-				const { start, end } = record.props
-				const nStart = { ...start }
-				const nEnd = { ...end }
-				if (nStart.type === 'binding') {
-					if (!nStart.isPrecise) {
-						nStart.normalizedAnchor = { x: 0.5, y: 0.5 }
-					}
-					delete nStart.isPrecise
-				}
-				if (nEnd.type === 'binding') {
-					if (!nEnd.isPrecise) {
-						nEnd.normalizedAnchor = { x: 0.5, y: 0.5 }
-					}
-					delete nEnd.isPrecise
-				}
-				return {
-					...record,
-					props: {
-						...record.props,
-						start: nStart,
-						end: nEnd,
-					},
-				}
-			},
-		},
-
-		[ArrowMigrationVersions.AddLabelPosition]: {
-			up: (record) => {
-				return {
-					...record,
-					props: {
-						...record.props,
-						labelPosition: 0.5,
-					},
-				}
-			},
-			down: (record) => {
-				const { labelPosition: _, ...props } = record.props
-				return {
-					...record,
-					props,
-				}
-			},
-		},
-	},
-})
