+++ packages/tlschema/src/shapes/TLEmbedShape.ts
@@ -1,4 +1,3 @@
-import { defineMigrations } from '@tldraw/store'
 import { T } from '@tldraw/validate'
 import { ShapePropsType, TLBaseShape } from './TLBaseShape'
 
@@ -626,130 +625,3 @@ export type EmbedDefinition = {
 	readonly toEmbedUrl: (url: string) => string | undefined
 	readonly fromEmbedUrl: (url: string) => string | undefined
 }
-
-const Versions = {
-	GenOriginalUrlInEmbed: 1,
-	RemoveDoesResize: 2,
-	RemoveTmpOldUrl: 3,
-	RemovePermissionOverrides: 4,
-} as const
-
-/** @internal */
-// eslint-disable-next-line deprecation/deprecation
-export const embedShapeMigrations = defineMigrations({
-	currentVersion: Versions.RemovePermissionOverrides,
-	migrators: {
-		[Versions.GenOriginalUrlInEmbed]: {
-			// add tmpOldUrl property
-			up: (shape) => {
-				const url = shape.props.url
-				const host = new URL(url).host.replace('www.', '')
-				let originalUrl
-				for (const localEmbedDef of EMBED_DEFINITIONS) {
-					if ((localEmbedDef as EmbedDefinition).hostnames.includes(host)) {
-						try {
-							originalUrl = localEmbedDef.fromEmbedUrl(url)
-						} catch (err) {
-							console.warn(err)
-						}
-					}
-				}
-
-				return {
-					...shape,
-					props: {
-						...shape.props,
-						tmpOldUrl: shape.props.url,
-						url: originalUrl ?? '',
-					},
-				}
-			},
-			// remove tmpOldUrl property
-			down: (shape) => {
-				let newUrl = shape.props.tmpOldUrl
-				if (!newUrl || newUrl === '') {
-					const url = shape.props.url
-					const host = new URL(url).host.replace('www.', '')
-
-					for (const localEmbedDef of EMBED_DEFINITIONS) {
-						if ((localEmbedDef as EmbedDefinition).hostnames.includes(host)) {
-							try {
-								newUrl = localEmbedDef.toEmbedUrl(url)
-							} catch (err) {
-								console.warn(err)
-							}
-						}
-					}
-				}
-
-				// eslint-disable-next-line @typescript-eslint/no-unused-vars
-				const { tmpOldUrl, ...props } = shape.props
-				return {
-					...shape,
-					props: {
-						...props,
-						url: newUrl ?? '',
-					},
-				}
-			},
-		},
-		[Versions.RemoveDoesResize]: {
-			up: (shape) => {
-				const { doesResize: _, ...props } = shape.props
-				return {
-					...shape,
-					props: {
-						...props,
-					},
-				}
-			},
-			down: (shape) => {
-				return {
-					...shape,
-					props: {
-						...shape.props,
-						doesResize: true,
-					},
-				}
-			},
-		},
-		[Versions.RemoveTmpOldUrl]: {
-			up: (shape) => {
-				const { tmpOldUrl: _, ...props } = shape.props
-				return {
-					...shape,
-					props: {
-						...props,
-					},
-				}
-			},
-			down: (shape) => {
-				return {
-					...shape,
-					props: {
-						...shape.props,
-					},
-				}
-			},
-		},
-		[Versions.RemovePermissionOverrides]: {
-			up: (shape) => {
-				const { overridePermissions: _, ...props } = shape.props
-				return {
-					...shape,
-					props: {
-						...props,
-					},
-				}
-			},
-			down: (shape) => {
-				return {
-					...shape,
-					props: {
-						...shape.props,
-					},
-				}
-			},
-		},
-	},
-})
