+++ packages/tlschema/src/shapes/TLGroupShape.ts
@@ -1,4 +1,3 @@
-import { defineMigrations } from '@tldraw/store'
 import { ShapeProps, TLBaseShape } from './TLBaseShape'
 
 /** @public */
@@ -9,7 +8,3 @@ export type TLGroupShape = TLBaseShape<'group', TLGroupShapeProps>
 
 /** @internal */
 export const groupShapeProps: ShapeProps<TLGroupShape> = {}
-
-/** @internal */
-// eslint-disable-next-line deprecation/deprecation
-export const groupShapeMigrations = defineMigrations({})
