+++ packages/tlschema/src/shapes/TLImageShape.ts
@@ -1,4 +1,3 @@
-import { defineMigrations } from '@tldraw/store'
 import { T } from '@tldraw/validate'
 import { assetIdValidator } from '../assets/TLBaseAsset'
 import { vecModelValidator } from '../misc/geometry-types'
@@ -27,45 +26,3 @@ export type TLImageShapeProps = ShapePropsType<typeof imageShapeProps>
 
 /** @public */
 export type TLImageShape = TLBaseShape<'image', TLImageShapeProps>
-
-const Versions = {
-	AddUrlProp: 1,
-	AddCropProp: 2,
-	MakeUrlsValid: 3,
-} as const
-
-/** @internal */
-// eslint-disable-next-line deprecation/deprecation
-export const imageShapeMigrations = defineMigrations({
-	currentVersion: Versions.MakeUrlsValid,
-	migrators: {
-		[Versions.AddUrlProp]: {
-			up: (shape) => {
-				return { ...shape, props: { ...shape.props, url: '' } }
-			},
-			down: (shape) => {
-				const { url: _, ...props } = shape.props
-				return { ...shape, props }
-			},
-		},
-		[Versions.AddCropProp]: {
-			up: (shape) => {
-				return { ...shape, props: { ...shape.props, crop: null } }
-			},
-			down: (shape) => {
-				const { crop: _, ...props } = shape.props
-				return { ...shape, props }
-			},
-		},
-		[Versions.MakeUrlsValid]: {
-			up: (shape) => {
-				const url = shape.props.url
-				if (url !== '' && !T.linkUrl.isValid(shape.props.url)) {
-					return { ...shape, props: { ...shape.props, url: '' } }
-				}
-				return shape
-			},
-			down: (shape) => shape,
-		},
-	},
-})
