+++ packages/tlschema/src/shapes/TLTextShape.ts
@@ -1,4 +1,3 @@
-import { defineMigrations } from '@tldraw/store'
 import { T } from '@tldraw/validate'
 import { DefaultColorStyle } from '../styles/TLColorStyle'
 import { DefaultFontStyle } from '../styles/TLFontStyle'
@@ -23,34 +22,3 @@ export type TLTextShapeProps = ShapePropsType<typeof textShapeProps>
 
 /** @public */
 export type TLTextShape = TLBaseShape<'text', TLTextShapeProps>
-
-const Versions = {
-	RemoveJustify: 1,
-} as const
-
-/** @internal */
-// eslint-disable-next-line deprecation/deprecation
-export const textShapeMigrations = defineMigrations({
-	currentVersion: Versions.RemoveJustify,
-	migrators: {
-		[Versions.RemoveJustify]: {
-			up: (shape) => {
-				let newAlign = shape.props.align
-				if (newAlign === 'justify') {
-					newAlign = 'start'
-				}
-
-				return {
-					...shape,
-					props: {
-						...shape.props,
-						align: newAlign,
-					},
-				}
-			},
-			down: (shape) => {
-				return { ...shape }
-			},
-		},
-	},
-})
